requirements.txt
amqp==2.5.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
coverage==4.5.3
Django==2.2.3
djangorestframework==3.10.1
factory-boy==2.12.0
Faker==2.0.0
jsonschema==3.0.1
kombu==4.6.3
Pillow==6.1.0
pkg-resources==0.0.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
vine==1.3.0
